hras_backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.11
pymysql==1.1.2
python-dotenv==1.0.0
google-generativeai==0.8.3